--- libp2p/pubsub/typing.py
@@ -1,21 +1,0 @@
-from collections.abc import (
-    Awaitable,
-)
-from typing import (
-    Callable,
-    Union,
-)
-
-from libp2p.peer.id import (
-    ID,
-)
-
-from .pb import (
-    rpc_pb2,
-)
-
-SyncValidatorFn = Callable[[ID, rpc_pb2.Message], bool]
-AsyncValidatorFn = Callable[[ID, rpc_pb2.Message], Awaitable[bool]]
-ValidatorFn = Union[SyncValidatorFn, AsyncValidatorFn]
-
-UnsubscribeFn = Callable[[], Awaitable[None]]
