+++ libp2p/tools/factories.py
@@ -37,7 +37,9 @@ from libp2p.crypto.keys import (
 )
 from libp2p.crypto.secp256k1 import create_new_key_pair as create_secp256k1_key_pair
 from libp2p.custom_types import (
+    TMuxerOptions,
     TProtocol,
+    TSecurityOptions,
 )
 from libp2p.host.basic_host import (
     BasicHost,
@@ -106,10 +108,6 @@ from libp2p.tools.constants import (
 )
 from libp2p.transport.tcp.tcp import (
     TCP,
-)
-from libp2p.transport.typing import (
-    TMuxerOptions,
-    TSecurityOptions,
 )
 from libp2p.transport.upgrader import (
     TransportUpgrader,
